Rearrange gems outside development and test groups in Gemfile
This commit modifies the Gemfile to move 'rspec-rails', 'simplecov', and 'capybara' out of the development and test groups. Now these gems are available for all environments, not only for development and testing. This change was necessary because these resources need to be accessible in production for bug detection and coverage analysis.